fix(settings): share font family helpers for mirrored XTerm
Move code-font helpers into lib/shared so XTermViewport stays
byte-identical across GUI and WebUI, unblocking the mirror check.
