+++ crates/agent-gui/src/components/project-tools/XTermViewport.tsx
@@ -3,8 +3,8 @@ import "@xterm/xterm/css/xterm.css";
 import { FitAddon } from "@xterm/addon-fit";
 import { Terminal as XTerm } from "@xterm/xterm";
 import { type CSSProperties, useEffect, useRef } from "react";
+import { CODE_FONT_FAMILY_CHANGE_EVENT, getCodeFontFamily } from "../../lib/shared/fontFamily";
 import { cn } from "../../lib/shared/utils";
-import { CODE_FONT_FAMILY_CHANGED_EVENT, resolveCodeFontFamily } from "../../lib/system/fontFamily";
 import type {
   TerminalClient,
   TerminalSession,
@@ -159,9 +159,7 @@ export function XTermViewport({
       cursorStyle: "block",
       cursorInactiveStyle: "outline",
       disableStdin: true,
-      fontFamily: resolveCodeFontFamily(
-        getComputedStyle(document.documentElement).getPropertyValue("--code-font-family"),
-      ),
+      fontFamily: getCodeFontFamily(),
       fontSize: 13,
       fontWeight: "normal",
       fontWeightBold: "bold",
@@ -207,16 +205,14 @@ export function XTermViewport({
         // xterm fit can throw while the panel is hidden or measuring at zero size.
       }
     };
-    fitAndResizeRef.current = fitAndResize;
 
-    const handleCodeFontFamilyChanged = (event: Event) => {
-      const codeFontFamily =
-        event instanceof CustomEvent ? event.detail?.codeFontFamily : undefined;
+    const handleCodeFontFamilyChange = (event: Event) => {
+      const codeFontFamily = (event as CustomEvent<string>).detail;
       if (typeof codeFontFamily !== "string") return;
       term.options.fontFamily = codeFontFamily;
       window.setTimeout(fitAndResize, 0);
     };
-    window.addEventListener(CODE_FONT_FAMILY_CHANGED_EVENT, handleCodeFontFamilyChanged);
+    window.addEventListener(CODE_FONT_FAMILY_CHANGE_EVENT, handleCodeFontFamilyChange);
 
     const resizeObserver = new ResizeObserver(() => {
       if (resizeTimerRef.current !== null) {
@@ -504,7 +500,7 @@ export function XTermViewport({
       streamInputUnsubscribe?.();
       streamHandle?.dispose();
       term.dispose();
-      window.removeEventListener(CODE_FONT_FAMILY_CHANGED_EVENT, handleCodeFontFamilyChanged);
+      window.removeEventListener(CODE_FONT_FAMILY_CHANGE_EVENT, handleCodeFontFamilyChange);
     };
   }, [client, session.id, session.projectPathKey]);
